requirements.txt
asgiref==3.11.1
Django==4.2.29
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
django-cors-headers==4.3.1
